frontend/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.0.0
motor>=3.3.0
pymongo>=4.6.0
certifi>=2023.0.0

